switch from cpp to c
