+++ src/wasmMath/add.h
@@ -1,14 +1,11 @@
-#ifndef DIVIDE_H
-#define DIVIDE_H
+#ifndef ADD_H
+#define ADD_H
 
 #ifdef __EMSCRIPTEN__
 #include <emscripten.h>
 EMSCRIPTEN_KEEPALIVE
 #endif
 
-extern "C" {
-
-int divide(int a, int b);
-}
+int add(int a, int b);
 
 #endif
