+++ src/wasmMath/divide.c
@@ -1,6 +1,3 @@
-#include "divide.hpp"
-
-extern "C" {
+#include "divide.h"
 
 int divide(int a, int b) { return a / b; }
-}
