+++ src/wasmMath/divide.h
@@ -1,14 +1,11 @@
-#ifndef ADD_H
-#define ADD_H
+#ifndef DIVIDE_H
+#define DIVIDE_H
 
 #ifdef __EMSCRIPTEN__
 #include <emscripten.h>
 EMSCRIPTEN_KEEPALIVE
 #endif
 
-extern "C" {
-
-int add(int a, int b);
-}
+int divide(int a, int b);
 
 #endif
