+++ src/wasmMath/multiply.c
@@ -1,6 +1,3 @@
-#include "multiply.hpp"
-
-extern "C" {
+#include "multiply.h"
 
 int multiply(int a, int b) { return a * b; }
-}
