+++ src/wasmMath/subtract.c
@@ -1,6 +1,3 @@
-#include "subtract.hpp"
-
-extern "C" {
+#include "subtract.h"
 
 int subtract(int a, int b) { return a - b; }
-}
